third_party/blink/web_tests/platform/generic/http/tests/filesystem/resolve-uri-expected.txt

Tests using webkitResolveLocalFileSystemURL to obtain an Entry from a URL

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

* Resolving a generated URL.
PASS expectedPath is actualPath
PASS isFile is true
* Resolving test file by hand
PASS expectedPath is actualPath
PASS isFile is true
* Resolving a URL with the wrong security origin (domain)
PASS "SecurityError" is errorName
* Resolving a URL with the wrong security origin (port)
PASS "SecurityError" is errorName
* Resolving a URL with the wrong security origin (scheme)
PASS "SecurityError" is errorName
* Resolving a completely bogus URL.
PASS "EncodingError" is errorName
* Resolving a URL with the wrong protocol
PASS "EncodingError" is errorName
* Resolving a URL with no slash between type and file
PASS "EncodingError" is errorName
* Resolving a URL with no slash between protocol and type (bogus port)
PASS "SecurityError" is errorName
* Resolve a path using backslashes
PASS expectedPath is actualPath
PASS isFile is true
* Resolve a directory
PASS expectedPath is actualPath
PASS isDirectory is true
* Resolve a path using a trailing slash
PASS expectedPath is actualPath
PASS isDirectory is true
* Resolving a persistent URL.
PASS expectedPath is actualPath
PASS isDirectory is true
Finished running tests.
PASS successfullyParsed is true

TEST COMPLETE
